requirements.txt
albumentations==1.3.1
ffmpeg==1.4
ffmpeg_python==0.2.0
gdown==4.7.1
matplotlib==3.7.1
numpy==1.24.1
opencv_python==4.8.0.74
opencv_python_headless==4.8.0.74
pandas==2.0.3
Pillow
scikit_learn==1.3.0
striprtf==0.0.25
thop==0.1.1.post2209072238
torch==2.0.1+cu118
torchmetrics==1.0.0
torchvision==0.15.2+cu118
tqdm==4.65.0
transformers==4.30.2
ultralytics==8.0.126
